% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateCMs.R
\name{aggregateCMs}
\alias{aggregateCMs}
\title{Aggregate dataframe with connectivity matrices according to new dates, spawning areas, nursery areas}
\usage{
aggregateCMs(dfrCMs, newDates = NULL, newSAs = NULL, newNAs = NULL,
  debug = FALSE)
}
\arguments{
\item{dfrCMs}{- dataframe with connectivity matrices, with numbers released and finished}

\item{newDates}{- vector of reclassified dates to aggregate to (or NULL to keep originals)}

\item{newSAs}{- vector of reclassified spawning areas to aggregate to (or NULL to keep originals)}

\item{newNAs}{- vector of reclassified nursery areas to aggregate to (or NULL to keep originals)}

\item{debug}{- flag to turn on debugging output}
}
\value{
dataframe with same column names as dfrCMs, but aggregated to new levels
}
\description{
Function to aggregate dataframe with connectivity matrices according to new dates, spawning areas, nursery areas.
}
\details{
input dataframe ("dfrCMs") is assumed to have columns in the order 
(date, spawning area, nursery area, numRel, numFin, and prFin)
although the names associated with these columns are arbitrary.
}

