% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregateNumBySrcSink.R
\name{aggregateNumBySrcSink}
\alias{aggregateNumBySrcSink}
\title{Aggregate numbers successful .}
\usage{
aggregateNumBySrcSink(dfrNbySS, factors = c("date", "start_alongshorezone",
  "end_alongshorezone", "totRel"), vars = c("totSet"))
}
\arguments{
\item{dfrNbySS}{- dataframe with numbers by date and source/sink pairs (e.g., from calcNumBySrcSink())}

\item{factors}{- factors to aggregate by}

\item{vars}{- variable(s) to aggregate}
}
\value{
dataframe with vars aggregated by levels of "factor" columns
}
\description{
Function to aggregate dataframe according to "factor" levels.
}
\details{

}

