% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAvgCM.R
\name{calcAvgCM}
\alias{calcAvgCM}
\title{Calculate an "average" connectivity matrix}
\usage{
calcAvgCM(allRes)
}
\arguments{
\item{allRes}{- the results list from a connectivity analysis}
}
\value{
list with 5 elements:
\itemize{
 \item avg  - the average connectivity matrix 
 \item stdv - the standard deviations in connectivity matrix form
 \item devs - the deviations in connectivity matrix form
 \item dmin - the minimum value in the devs
 \item dmax - the maximum value in the devs
}
}
\description{
Calculates an "average" connectivity matrix from 
results of a connectivity analysis.
}
\details{
none.
}
