% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCM.R
\name{calcConnectivityMatrix}
\alias{calcConnectivityMatrix}
\title{Calculate a connectivity matrix for one DisMELS model run.}
\usage{
calcConnectivityMatrix(ibmResTbl = NULL, cellsTbl = NULL,
  lhsTypeInfo = getLifeStageInfo.ATF(),
  spawningZones = c("SpawningArea_300to600m"),
  nurseryZones = c("NurseryArea_000to050m", "NurseryArea_050to150m"),
  writeCSVs = TRUE, outNumRel = "NumbersReleasedFromSpawningAreas.csv",
  outConn = "ConnectivityToNurseryAreasFromSpawningAreas.csv",
  outIndivs = "IndivStartEndPositions.csv")
}
\arguments{
\item{ibmResTbl}{- name of DisMELS connectivity results file or resulting dataframe (can be NULL)}

\item{cellsTbl}{- name of classified grid cells file or resulting dataframe (can be NULL)}

\item{lhsTypeInfo}{- list object with life stage info}

\item{spawningZones}{- vector of names of zones used as spawning areas in the IBM}

\item{nurseryZones}{- vector of names of zones used as nursery areas in the IBM}

\item{writeCSVs}{- flag (T/F) to write output files}

\item{outNumRel}{- csv filename for initial numbers released from spawning areas}

\item{outConn}{- csv filename for connectivity to nursery areas from spawning areas}

\item{outIndivs}{- csv filename for individual start and end locations}
}
\value{
list w/ 2 elements:\cr
dfrCM - dataframe with probability of transport to a sink (nursery area) conditioned on starting in a source (spawning area)\cr
dfrNR - numbers by source (spawning area)\cr
dfrIC - dataframe of start and end locations for each individual in the model run\cr
}
\description{
Function to calculate a connectivity matrix based on one DisMELS model run.
}
\details{
If the ibmResTbl dataframe is NULL, the user can select the file using a 
file dialog box).\cr \cr

If the cellsTbl dataframe is NULL, the user can select the file using a 
file dialog box).\cr \cr

If the writeCSVs flag is set, the two files specified by outConn and outIndivs are written before the function is exited.
}

