% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIC.R
\name{calcIC}
\alias{calcIC}
\title{Calculate individual connectivity for one DisMELS model run}
\usage{
calcIC(ibmResTbl = NULL, cellsTbl = NULL,
  lhsTypeInfo = getLifeStageInfo.ATF(),
  spawningZones = c("SpawningArea_300to600m"), writeCSV = FALSE,
  outIndivs = "IndivStartEndPositions.csv")
}
\arguments{
\item{ibmResTbl}{- name of DisMELS connectivity results file or resulting dataframe (can be NULL)}

\item{cellsTbl}{- name of classified grid cells file or resulting dataframe (can be NULL)}

\item{lhsTypeInfo}{- list object with life stage info}

\item{spawningZones}{- vector of names of zones used as spawning areas in the IBM}

\item{writeCSV}{- flag (T/F) to write output file}

\item{outIndivs}{- csv filename for individual connectivity (start and end locations)}
}
\value{
dataframe of start and end locations for each individual in the model run
}
\description{
Function to calculate individual connectivity based on one DisMELS model run.
}
\details{
If the ibmResTbl dataframe is NULL, the user can select the file using a 
file dialog box.\cr \cr

If the cellsTbl dataframe is NULL, the user can select the file using a 
file dialog box.\cr \cr

If the writeCSV flag is set, the inidividual connectivity data frame (dfrIC) 
is written to the file specified by 'outIndivs'.
}

