% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcLinRegs.RbyI.R
\name{calcLinRegs.RbyI}
\alias{calcLinRegs.RbyI}
\title{Calculate linear regressions between recruitment and index time series}
\usage{
calcLinRegs.RbyI(mdfrZR, mdfrI, ylab, vars, nrows = 2, labelByGroup = FALSE)
}
\arguments{
\item{mdfrZR}{- melted dataframe with standardized recruitment time series}

\item{mdfrI}{- melted dataframe with raw index time series}

\item{ylab}{- label for y axis in time series plot of raw time series}

\item{nrows}{- number of rows for LR plots}

\item{labelByGroup}{- flag to use "group" column to organize linear regression analysis}
}
\value{
list with elements
* mdfrZIs - dataframe with z-scores for index time series
* lms     - list with linear model (lm) results by index
* summary - dataframe with summaries of linear model results
* plots   - list of ggplot2 objects (raw time series, standardized time series, fits to lm's)
}
\description{
Function to calculate linear regressions between recruitment and index time series
}
\details{

}

