% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLinRegs.YsByX.R
\name{calcLinRegs.YsByX}
\alias{calcLinRegs.YsByX}
\title{Calculate linear regressions for time series in Y as function of time series in X.}
\usage{
calcLinRegs.YsByX(mdfrX, mdfrY, yvars = NULL, ylab = NULL, nrows = 2,
  coord_fixed = FALSE, labelByGroup = FALSE)
}
\arguments{
\item{mdfrX}{- melted dataframe with independent (X) time series}

\item{mdfrY}{- melted dataframe with dependent (Y) time series}

\item{yvars}{- vector of dependent variables to process (or NULL for all)}

\item{ylab}{- label associated with  dependent variables}

\item{nrows}{- number of rows for LR plots}

\item{coord_fixed}{- flag to use equal x/y unit dimensions in LR plots}

\item{labelByGroup}{- flag to use "group" column to organize linear regression analysis}
}
\value{
list of named sublists by unique X with elements
* lms     - list with linear model (lm) results by Y for X
* summary - dataframe with summaries of linear model results
* plots   - list of ggplot2 objects (p1=standardized time series, p2=fits to lm's)
}
\description{
Function to calculate linear regressions between time series in Y as function of time series in X.
}

