% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLRs.RbyI.R
\name{calcLRs.RbyI}
\alias{calcLRs.RbyI}
\title{Calculate linear regressions between recruitment and index time series}
\usage{
calcLRs.RbyI(mdfrZR, mdfrI, ylab, vars, nrows = 2, labelByGroup = FALSE)
}
\arguments{
\item{mdfrZR}{- melted dataframe with standardized recruitment time series}

\item{mdfrI}{- melted dataframe with raw index time series}

\item{ylab}{- label for y axis in time series plot of raw time series}

\item{vars}{- subset of index time series to include in linear regresssion analysis}

\item{nrows}{- number of rows for LR plots}

\item{labelByGroup}{- flag to use "group" column to organize linear regression analysis}
}
\value{
list with elements
\itemize{
 \item mdfrZIs - dataframe with z-scores for index time series
 \item lms     - list with linear model (lm) results by index
 \item summary - dataframe with summaries of linear model results
 \item plots   - list of ggplot2 objects (raw time series, standardized time series, fits to lm's)
}
}
\description{
Function to calculate linear regressions between recruitment and index time series
}

