% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLRs.YbyXs.R
\name{calcLRs.YbyXs}
\alias{calcLRs.YbyXs}
\title{Calculate linear regressions for time series in Y as function of time series in X.}
\usage{
calcLRs.YbyXs(mdfrX, mdfrY, yvars = NULL, xlab = "index", ylab = "index",
  nrows = 2, coord_fixed = FALSE, labelByGroup = FALSE, verbose = FALSE)
}
\arguments{
\item{mdfrX}{- melted dataframe with independent (X) time series}

\item{mdfrY}{- melted dataframe with dependent (Y) time series}

\item{yvars}{- vector of dependent variables to process (or NULL for all)}

\item{xlab}{- label associated with independent variables}

\item{ylab}{- label associated with dependent variables}

\item{nrows}{- number of rows for LR plots}

\item{coord_fixed}{- flag to use equal x/y unit dimensions in LR plots}

\item{labelByGroup}{- flag to use "group" column to organize linear regression analysis}

\item{verbose}{- flag to print messages}
}
\value{
list with named elements 'res', 'summary' and 'plot': \cr
\itemize{
 \item res - a list with sublists by unique Y with elements
 \itemize{
   \item lms     - list with linear model (lm) results by Y for X
   \item summary - dataframe with summaries of linear model results
   \item plots   - list of ggplot2 objects (p1=standardized time series, p2=fits to lm's)
 }
 \item summary - a dataframe summarizing all LR results. Columns are
 \itemize{
   \item ygroup - dependent variable group label
   \item y      - dependent variable label
   \item xgroup - independent variable group label
   \item x      - independent variable label
   \item n      - number of valid data points
   \item rho    - Pearson's correlation coefficient
   \item rsq    - R-squared for linear fit
   \item p      - P-value (uncorrected for multiple comparisons)
 }
 \item plot - a ggplot object.
 \item data - dataframe with data involved in linear models
}
}
\description{
Function to calculate linear regressions between time series in Y as function of time series in X.
}

