% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLRs.YsByXs.R
\name{calcLRs.YsByXs}
\alias{calcLRs.YsByXs}
\title{Calculate linear regressions for time series in Y as function of time series in X}
\usage{
calcLRs.YsByXs(mdfrX, mdfrY, verbose = FALSE)
}
\arguments{
\item{mdfrX}{- melted dataframe with independent (X) time series}

\item{mdfrY}{- melted dataframe with dependent (Y) time series}

\item{verbose}{- flag to print messages}
}
\value{
list with named elements 'res', 'summary' and 'plot': \cr
\itemize{
 \item res - a list with sublists by unique YG:YV:XG:XV with elements
 \itemize{
   \item lms     - list with linear model (lm) results by Y for X
   \item summary - dataframe with summaries of linear model results
 }
 \item summary - a dataframe summarizing all LR results. Columns are
 \itemize{
   \item ygroup - dependent variable group label
   \item y      - dependent variable label
   \item xgroup - independent variable group label
   \item x      - independent variable label
   \item n      - number of valid data points
   \item rho    - Pearson's correlation coefficient
   \item rsq    - R^2 for linear fit
   \item adj.rsq - adjusted R^2 for linear fit
   \item F      - F statistic
   \item prF    - Pr(>F) (p-value uncorrected for multiple comparisons)
   \item aicc   - AICc value
   \item del.aicc - AICc relative to null model [i.e., AICc(model) - AICc(null)]
 }
 \item data - dataframe with data involved in linear models
 \itemize{
   \item ygroup - dependent variable group label
   \item yvar   - dependent variable label
   \item xgroup - independent variable group label
   \item xvar   - independent variable label
   \item year   - year
   \item x      - independent value in LR
   \item y      - dependent value in LR
 }
}
}
\description{
Function to calculate linear regressions between time series in Y as function of time series in X.
}
\details{
Requires packages 'plyr', 'reshape2', and 'glmulti'. Datasets are standardized as z-scores, 
so the constant term in the linear model is dropped.


none.
}
