% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcNumBySrcSink.R
\name{calcNumBySrcSink}
\alias{calcNumBySrcSink}
\title{Calculate numbers of successful individuals by source, sink areas.}
\usage{
calcNumBySrcSink(dfrNumBySrc, dfrConMats)
}
\arguments{
\item{dfrNumBySrc}{}

\item{dfrConMats}{}
}
\description{
Function to calculate numbers of successful individuals by source, sink areas.
}
\details{

}

