% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOverlapCoefficient.CMs.R
\name{calcOverlapCoefficient.CMs}
\alias{calcOverlapCoefficient.CMs}
\title{UNFINISHED: Calculate the overlap coefficient (OC) time series for a set of connectivity matrices}
\usage{
calcOverlapCoefficient.CMs(dfrCMs)
}
\arguments{
\item{dfrCMs}{- dataframe with connectivity matrices from a connectivity analysis}
}
\value{
TODO: complete this
}
\description{
Calculates the overlap coefficient (OC) time series for a set of connectivity matrices.
}
\details{
None.
}

