% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatAnnualTables.R
\name{concatAnnualTables}
\alias{concatAnnualTables}
\title{Concatenate annual tables (connectivity matrices, individual connectivities, numbers by source), write to csv files.}
\usage{
concatAnnualTables(allRes = NULL, type = c("prbSinkGivenSource",
  "tblIndivConn", "numSource"), writeCSV = TRUE, folder = getwd(),
  basename = c("ConnMats", "IndivConns", "numsBySource"))
}
\arguments{
\item{allRes}{- the allRes list from running calcConnectivityMatrices(...)}

\item{type}{- type of table to concatenate}

\item{writeCSV}{- flag (T/F) to write concatenated table to csv}

\item{folder}{- folder to write csv file}

\item{basename}{- base name for output csv file}
}
\value{
dataframe with concatenated matrices
}

