% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatEIs.R
\name{concatEIs}
\alias{concatEIs}
\title{Concatenate files with environmental indices}
\usage{
concatEIs(caption = "Select environmental index files", types = "txt",
  hasHeader = FALSE, id.vars = "year", out.csv = "test.csv",
  out.dir = NULL)
}
\arguments{
\item{caption}{- caption for file dialog}

\item{types}{- vector of extension types to include in file dialog}

\item{hasHeader}{- flag indicating file has header}

\item{id.vars}{- vector of column names to use as id vars}

\item{out.csv}{- name of output csv file}

\item{out.dir}{- directory for output csv file}
}
\value{
melted dataframe with concatenated indices
}
\description{
Function to concatenate files with environmental indices.
}
\details{
Uses \code{tcltk::tk_choose_files}.
}

