% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLabelToExpr.R
\name{convertLabelToExpr}
\alias{convertLabelToExpr}
\title{Convert plot with newline characters to plotmath}
\usage{
convertLabelToExpr(txt, test = FALSE)
}
\arguments{
\item{txt}{- character vector of labels.}

\item{test}{- flag to plot labels (for debugging)}
}
\value{
character vector of labels
}
\description{
Function to convert plot with newline characters to plotmath
}
\details{
Converts the vector of labels to a text vector which
can be parsed to yield a plotmath expression. Each label is first split
by newline characters. The first split  is re-split using ":". The 
last line is split by "R^2". The resulting splits are recombined in nested
'atop(x,y)' functions (see plotmath), with the first re-split interpreted
as a plotmath expression, as is 'R^2'.

test=TRUE requires ggplot2 package.
}
