% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIndivs.MultipleRuns.R
\name{extractIndivs.MultipleRuns}
\alias{extractIndivs.MultipleRuns}
\title{Extracts and re-orders a series of DisMELS model results files by individual}
\usage{
extractIndivs.MultipleRuns(YYYYs = NULL, MMs = NULL, DDs = NULL,
  lhsTypeInfo = NULL, lhsTypes = NULL,
  successful = "where (end_typeName=\\"benthic.juvenile\\")",
  nurseryZones = c("NurseryArea_000to050m", "NurseryArea_050to150m"),
  onlySuccessful = TRUE, onlyUnsuccessful = FALSE,
  inpDir.IndivConn = NULL, basename.IndivConn = "IndivConn",
  inpDir.Results = NULL, basename.Results = "Results", outDir = NULL,
  basename.Out = "SuccessfulIndivs", verbose = FALSE)
}
\arguments{
\item{YYYYs}{- years to extract (can be NULL)}

\item{MMs}{- months to extract (can be NULL)}

\item{DDs}{- days to extract (can be NULL)}

\item{lhsTypeInfo}{- life stage info list for IBM}

\item{lhsTypes}{- lhs types to keep for output (NULL = all)}

\item{successful}{- sql "where" clause to determine final state for individuals regarded as successful}

\item{nurseryZones}{- vector of nursery area names}

\item{onlySuccessful}{- flag (T/F) to extract only successful indivs (if TRUE)}

\item{onlyUnsuccessful}{- flag (T/F) to extract only unsuccessful indivs (if TRUE)}

\item{inpDir.IndivConn}{- folder with input individual connectivity matrix files}

\item{basename.IndivConn}{- base name for individual connectivity matrix files}

\item{inpDir.Results}{- folder with input DisMELS model results files}

\item{basename.Results}{- base name for DisMELS model results files}

\item{outDir}{- output folder}

\item{basename.Out}{- base name for output files}

\item{verbose}{- flag (T/F) to print debugging info}
}
\value{
no return value
}
\description{
Function to extract and re-order a series of DisMELS model results by individual.
}
\details{
This function assumes files are named something like 'basenameYYYYMMDD.csv'. 
Note that at least one of YYYYs, MMs, and DDs must be a non-null vector.
}

