% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIndivs.R
\name{extractIndivs}
\alias{extractIndivs}
\title{Extract tracks of individuals from a DisMELS model run}
\usage{
extractIndivs(indivConn = NULL, results = NULL,
  lhsTypeInfo = getLifeStageInfo.ATF(), lhsTypes = NULL,
  onlySuccessful = FALSE, onlyUnsuccessful = FALSE,
  successful = "where (end_typeName=\\"benthic.juvenile\\")",
  nurseryZones = c("NurseryArea_000to050m", "NurseryArea_050to150m"),
  returnList = TRUE, writeOutput = TRUE, outDir = "./",
  outBaseCSV = "allIndivs", verbose = FALSE)
}
\arguments{
\item{indivConn}{- individual connectivity file (or dataframe)}

\item{results}{- DisMELS results file (or dataframe)}

\item{lhsTypeInfo}{- life stage info list for IBM}

\item{lhsTypes}{- vector of lhs types to process (NULL = all)}

\item{onlySuccessful}{- flag (T/F) to extract only successful indivs (if TRUE)}

\item{onlyUnsuccessful}{- flag (T/F) to extract only unsuccessful indivs (if TRUE)}

\item{successful}{- sql "where" clause to determine final state for individuals regarded as successful}

\item{nurseryZones}{- vector of names of zones used as nursery areas in the IBM}

\item{returnList}{- flag (T/F) indicating whether a list should be returned (otherwise NULL--helps with memory)}

\item{writeOutput}{- flag (T/F) indicating whether output files should be written}

\item{outDir}{- folder for output}

\item{outBaseCSV}{- base name for output csv files}

\item{verbose}{- flag (T/F) to print debugging info}
}
\value{
list of dataframes with indiv results by lhs type name if returnList=TRUE, otherwise NULL.
}
\description{
Function to extract tracks of individuals from a DisMELS model run.
}
\details{
If indivConn or results are NULL, the user will be prompted to select the
corresponding file(s) using a file chooser dialog.
Uses functions \code{wtsUtilities::getCSV} and \code{sqldf::sqldf}.
}

