% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResultsByLHS.MultipleRuns.R
\name{extractResultsByLHS.MultipleRuns}
\alias{extractResultsByLHS.MultipleRuns}
\title{Extracts and re-exports a series of DisMELS model results files by life stage.}
\usage{
extractResultsByLHS.MultipleRuns(YYYYs = NULL, MMs = NULL, DDs = NULL,
  lhsTypeInfo = NULL, lhsTypes = NULL, inpDir.Results = NULL,
  basename.Results = "Results", outDir = NULL, basename.Out = "Results")
}
\arguments{
\item{YYYYs}{- years to extract (can be NULL)}

\item{MMs}{- months to extract (can be NULL)}

\item{DDs}{- days to extract (can be NULL)}

\item{lhsTypeInfo}{- life stage info list for IBM}

\item{lhsTypes}{- lhs types to keep for output (NULL = all)}

\item{inpDir.Results}{- folder with input DisMELS model results files}

\item{basename.Results}{- base name for DisMELS model results files}

\item{outDir}{- output folder}

\item{basename.Out}{- base name for output files}
}
\value{
no return value
}
\description{
Function to extract and re-export a series of DisMELS model results by life stage.
}
\details{
This function assumes files are named something like 'basenameYYYYMMDD.csv'. 
Note that at least one of YYYYs, MMs, and DDs must be a non-null vector.
}

