% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractResultsByLHS.R
\name{extractResultsByLHS}
\alias{extractResultsByLHS}
\title{Extract results from DisMELS model run to individual files by life stage.}
\usage{
extractResultsByLHS(results, lhsTypeInfo = getLifeStageInfo.ATF(),
  lhsTypes = NULL, returnList = TRUE, writeOutput = TRUE,
  outDir = dirname(results), outBaseCSV = "Results")
}
\arguments{
\item{results}{- filename for results}

\item{lhsTypeInfo}{- list with life stage info for the model that created the results}

\item{lhsTypes}{- life stage types (names) to process}

\item{returnList}{- flag to return list of output (does nothing now)}

\item{writeOutput}{- flag to write output to file}

\item{outDir}{- folder to write output csv files}

\item{outBaseCSV}{- base name for output csv files}
}
\description{
Function to extract results from DisMELS model run to individual files by life stage.
}
\details{

}

