% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCM.dataframe.R
\name{plotCM.dataframe}
\alias{plotCM.dataframe}
\title{Plot a connectivity matrix from a dataframe}
\usage{
plotCM.dataframe(tbl, zmax = NULL, zmin = 0, title = NULL, xlab = xlab,
  ylab = ylab, devtype = "win", cex = 1 * (devtype == "win") + 1.5 *
  (devtype != "win"), colorscale = c("hot", "cold", "coldhot"),
  plotfile = "CM.png", plotsize = c(970, 780))
}
\arguments{
\item{tbl}{- dataframe or matrix representation of a connectivity matrix (can be NULL)}

\item{zmax}{- max value for color scale}

\item{zmin}{- min value for color scale}

\item{title}{- title for plot}

\item{xlab}{- label for x axis}

\item{ylab}{- label for y axis}

\item{devtype}{- device type ('win' or 'png')}

\item{cex}{- character expansion factor for plot}

\item{colorscale}{- color scale to use for plot ('hot', 'cold' or 'coldhot')}

\item{plotfile}{- file to save plot to (if devtype != 'win')}

\item{plotsize}{- 2 element vector giving plot size in pixels (c(W,H))}
}
\description{
Plot a connectivity matrix from a dataframe.
}
\details{
deprecated in favor of \code{plotCM.GG}.
}

