% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCM.R
\name{plotCM}
\alias{plotCM}
\title{Plot a connectivity matrix}
\usage{
plotCM(tbl = NULL, csvfile = NULL, zmax = NULL, zmin = 0,
  title = NULL, xlab = "Nursery Areas/Alongshore Zones",
  ylab = "Spawning Areas/Alongshore Zones", devtype = "win", cex = 1 *
  (devtype == "win") + 1.5 * (devtype != "win"), colorscale = c("hot", "cold",
  "coldhot"), plotfile = "CM.png", plotsize = c(970, 780))
}
\arguments{
\item{tbl}{- dataframe or matrix representation of a connectivity matrix (can be NULL)}

\item{csvfile}{- csv file to read connectivity matrix from (can be NULL)}

\item{zmax}{- max value for color scale}

\item{zmin}{- min value for color scale}

\item{title}{- title for plot}

\item{xlab}{- label for x axis}

\item{ylab}{- label for y axis}

\item{devtype}{- 'win' or 'png'}

\item{cex}{- character expansion factor for plot}

\item{colorscale}{- color scale to use for plot ('hot', 'cold' or 'coldhot')}

\item{plotfile}{- file to save plot to (if devtype != 'win')}

\item{plotsize}{- 2 element vector giving plot size in pixels (c(W,H))}
}
\description{
Plot a connectivity matrix (from a matrix, dataframe, or csv file).
}
\details{
- deprecated in favor of \code{plotCM.GG}. Calls \code{plotCM.matrix} or \code{plotCM.dataframe},
depending on class of 'tbl'.
}
