% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotConnectivityMatrixGG.R
\name{plotConnectivityMatrixGG}
\alias{plotConnectivityMatrixGG}
\title{Plot connectivity matrices using ggplot2.}
\usage{
plotConnectivityMatrixGG(dfr, x = "nursery area", y = "spawning area",
  z = "Pr(N|S)", xLab = x, yLab = y, zLab = z, xMin = NULL,
  xMax = NULL, yMin = NULL, yMax = NULL, zMin = NULL, zMax = NULL,
  facetWrap = NULL, facetGrid = NULL, ncol = 4, reverseX = FALSE,
  reverseY = FALSE, useGrad2 = FALSE)
}
\arguments{
\item{dfr}{- dataframe to plot}
}
\description{
Function to plot connectivity matrices using ggplot2.
}
\details{

}

