% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotConnInfoByArea.R
\name{plotConnInfoByArea}
\alias{plotConnInfoByArea}
\title{Plot dataframe with connectivity info aggregated over sources or sinks as time series and boxplots.}
\usage{
plotConnInfoByArea(dfr, time = "year", area = "spawning area",
  y = "Pr(.|S)", timeLab = time, areaLab = area, yLab = y, facets = EW
  ~ .)
}
\arguments{
\item{dfr}{- dataframe with conncetivity info aggregated over sources (spawning areas) or sinks (nursery areas)}

\item{time}{- name of column associated with times}

\item{area}{- name fo column associated with areas}

\item{y}{- name of column to plot as y axis}

\item{timeLab}{- label for times}

\item{areaLab}{- label for areas}

\item{yLab}{- label for y axis}

\item{facets}{- formula to facet by}
}
\value{
list with ggplot2 elements pTS (time series) and pBP (boxplot)
}
\description{
Function to plot dataframe with connectivity info aggregated over sources or sinks as time series and boxplots.
}
\details{

}

