% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotFractionRemaining.R
\name{plotFractionRemaining}
\alias{plotFractionRemaining}
\title{Plot fraction remaining by age-in-stage.}
\usage{
plotFractionRemaining(dfr = NULL, subsets = NULL,
  lhsTypeInfo = getLifeStageInfo.ATF(), nurseryAlongshoreZones = NULL,
  nurseryDepthZones = NULL, spawningAlongshoreZones = NULL,
  spawningDepthZones = NULL, title = NULL, ggtheme = theme_grey())
}
\arguments{
\item{dfr}{- dataframe or filename (or NULL) from call to extractIndivs(...)}

\item{lhsTypeInfo}{- life stage info list for IBM (req'd if onlySuccessful=TRUE or onlyUnsuccessful=TRUE)}

\item{nurseryAlongshoreZones}{- vector of id's for alongshore nursery zones to include}

\item{nurseryDepthZones}{- vector of names for depth zones used as nursery areas in the IBM (req'd if onlySuccessful=TRUE)}

\item{spawningAlongshoreZones}{- vector of id's for alongshore spawning zones to include}

\item{spawningDepthZones}{- vector of names for depth zones used as spawning areas in the IBM (req'd if onlySuccessful=TRUE)}

\item{title}{- plot title}

\item{ggtheme}{- ggplot2 theme for plot}

\item{subset}{- list with subsets of alongshore spawning zones to include in plots}
}
\description{

}

