% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistogram.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Plot a histogram for a column in a dataframe}
\usage{
plotHistogram(dfr, x, label = "", fill = "grey50", binwidth = NULL,
  lines = NULL, colour = "black", linetype = 2, linesize = 1.25,
  showPct = FALSE, reverse = FALSE, flip = TRUE, facets = NULL,
  showPlot = FALSE)
}
\arguments{
\item{dfr}{- dataframe}

\item{x}{- name of column for histogram}

\item{label}{- bin axis label}

\item{fill}{- histogram bin colour}

\item{binwidth}{- bin width for histogram}

\item{lines}{- vector of (nominally vertical) reference line positions}

\item{colour}{- colour for reference lines}

\item{linetype}{- type for reference lines}

\item{linesize}{- size for reference lines}

\item{showPct}{- flag to include \% observations falling outside lines as text on plot}

\item{reverse}{- reverse bin order}

\item{flip}{- flag to flip orientation}

\item{facets}{- expression for faceting using facet_grid()}

\item{showPlot}{- flag to print plot}
}
\value{
a ggplot2 object with an extra list element, 'xtra'. 'xtra' is a 
list with one element: pct.outside, which gives the \% observations falling
outside the range of the 'lines' (if specified), or NULL.
}
\description{
Function to plot a histogram for a column in a dataframe
}

