% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotTimeSeriesofIndivValues.R
\name{plotTimeSeriesofIndivValues}
\alias{plotTimeSeriesofIndivValues}
\title{Plot time series of a variable for individuals}
\usage{
plotTimeSeriesofIndivValues(dfr = NULL, vars = NULL, by = "age",
  lhsTypeInfo = getLifeStageInfo.ATF(), nurseryAlongshoreZones = NULL,
  nurseryDepthZones = NULL, spawningAlongshoreZones = NULL,
  spawningDepthZones = NULL, xlab = NULL, ylabs = NULL, titles = NULL,
  subsets = NULL, facets = NULL, ggtheme = theme_grey())
}
\arguments{
\item{dfr}{- dataframe or filename (or NULL) from call to extractIndivs(...)}

\item{vars}{- name(s) of variable(s) (columns) to plot}

\item{by}{- name of x-axis variable to plot by}

\item{lhsTypeInfo}{- life stage info list for IBM (req'd if onlySuccessful=TRUE or onlyUnsuccessful=TRUE)}

\item{nurseryAlongshoreZones}{- alongshore id's for nursery zones to include}

\item{nurseryDepthZones}{- vector of names of depth zones used as nursery areas in the IBM (req'd if onlySuccessful=TRUE)}

\item{spawningAlongshoreZones}{- alongshore id's for nursery zones to include}

\item{spawningDepthZones}{- vector of names of zones used as spawning areas in the IBM (req'd if onlySuccessful=TRUE)}

\item{xlab}{- x axis label}

\item{ylabs}{- y axis labels (character vector or list of expressions/character vectors)}

\item{titles}{- plot titles (character vector or list of expressions/character vectors)}

\item{facets}{- expression for creating facets (use bquote(...))}

\item{ggtheme}{- ggplot2 theme for plot}

\item{subset}{- list with subsets of alongshore spawning zones to include in plots}
}
\description{

}

