% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unzipDisMELSResults.R
\name{unzipDisMELSResults}
\alias{unzipDisMELSResults}
\title{Unzips and renames a series of DisMELS results or connectivity files.}
\usage{
unzipDisMELSResults(YYYYs = NULL, MMs = NULL, DDs = NULL,
  basename = "Results", dir.inp = getwd(), dir.out = getwd())
}
\arguments{
\item{YYYYs}{= years to extract (can be NULL)}

\item{MMs}{= months to extract (can be NULL)}

\item{DDs}{= days to extract (can be NULL)}

\item{basename}{=  base name for zipped/unzipped files}

\item{dir.inp}{= folder containing the zip files}

\item{dir.out}{= folder where unzipped files will be written}
}
\description{
Function to unzip and rename a series of DisMELS results or connectivity files.
}
\details{


This function assumes zip files are named something like 'basenameYYYYMMDD.csv.zip' and
that the associated zipped csv file is 'basename.csv'. When the csv file is unzipped
in the folder specified by dir.out, it will be renamed 'basenameYYYYMMDD.csv'. Note
that at least one of YYYYs, MMs, and DDs must be a non-null vector.
}

