\name{writeConnectivityMatrices.ToCSV}
\alias{writeConnectivityMatrices.ToCSV}
\title{Writes connectivity matrices to csv files.}
\usage{
writeConnectivityMatrices.ToCSV(allRes = NULL, concat = FALSE,
  folder = getwd(), basename = "ConnMat")
}
\arguments{
  \item{allRes}{- the allRes list from running
  calcConnectivityMatrices(...)}

  \item{concat}{- flag (T/F) to concatenate connectivity
  matrices into 1 output csv file}

  \item{folder}{- folder to write csv file(s) to}

  \item{basename}{- base name for output csv files}
}
\description{

}

