% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeIndivConns.R
\name{writeIndivConns}
\alias{writeIndivConns}
\title{Write dataframe with individual connectivity results to files by date}
\usage{
writeIndivConns(dfrICs, folder = getwd(), basename = "IndivConn")
}
\arguments{
\item{dfrICs}{- dataframe}

\item{folder}{- folder to write to}

\item{basename}{- base name for csv files}
}
\description{
Function to write dataframe with individual connectivity results to files by date.
}
\details{
none
}

