% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeIndividualConnectivity.ToCSV}
\alias{writeIndividualConnectivity.ToCSV}
\title{Write individual connectivity results to csv files.}
\usage{
writeIndividualConnectivity.ToCSV(allRes = NULL, concat = FALSE,
  folder = getwd(), basename = "IndivConn")
}
\arguments{
\item{allRes}{- the allRes list from running calcConnectivityMatrices(...)}

\item{concat}{- flag (T/F) to concatenate individual connectivity results into 1 output csv file}

\item{folder}{- folder to write csv file(s) to}

\item{basename}{- base name for output csv files}
}
\description{

}

