% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConnectivityMatrices.R
\name{calcConnectivityMatrices}
\alias{calcConnectivityMatrices}
\title{Calculate connectivity matrices for a set of DisMELS model runs.}
\usage{

  calcConnectivityMatrices(resDir = "C:\\\\Projects\\\\GOA_IERP\\\\IBM_Runs\\\\ATF\\\\FullSeries",
  connResBase = "ConnYYYYMMDD.csv", cellsTbl = file.path(resDir,
  "ATF_ClassifiedCGOAGridCells.csv"), years = as.character(1996:2011),
  months = c("01"), days = c("01"),
  spawningZones = c("SpawningArea_300to600m"),
  nurseryZones = c("NurseryArea_000to050m", "NurseryArea_050to150m"),
  lhsTypeInfo = getLifeStageInfo.ATF(), writeCSVs = TRUE,
  folder = getwd(), basename = "")
}
\arguments{
\item{resDir}{- path to base directory for DisMELS results files}

\item{connResBase}{- file name template for connectivity results files from DisMELS (e.g., "ConnYYYYMMDD.csv")}

\item{cellsTbl}{- name of classified grid cells file or dataframe based on classified grid cells file (can be NULL)}

\item{years}{- vector of years corresponding to model simulation names (YYYY)}

\item{months}{- vector of months corresponding to model simulation names (MM)}

\item{days}{- vector of days corresponding to model simulation names (DD)}

\item{spawningZones}{- vector of names of zones used as spawning areas in the IBM}

\item{nurseryZones}{- vector of names of zones used as nursery areas in the IBM}

\item{lhsTypeInfo}{- list object with life stage info}

\item{writeCSVs}{- flag (T/F) to write concatenated dataframes to csv}

\item{folder}{- folder to write csv files to}

\item{basename}{- base name for output csv files}
}
\value{
list w/ 3 elements:\cr
dfrCMs - dataframe with connectivity matrices results\cr
dfrICs - dataframe with individual connectivity results\cr
dfrNRs - dataframe with numbers released
}
\description{
Function to calculate a set of connectivity matrices corresponding to
a set of DisMELS model runs.
}
\details{
If the 'cellsTbl' is a filename,it will be read to create an associated dataframe. If cellsTbl is null, the user can select
the classified grid cells file using a file dialog.\cr\cr

Connectivity results files from the DisMELS model runs should be of the form "ConnYYYYMMDD.csv",
where "Conn" is arbitrary (but consistent among results files) and YYYY indicates the year (if applicable), MM the month (if applicable),
and DD the day (if applicable) for a particular model run. If YYYY, MM, or DD is **not** part of the file name, use "" as the 
function input for the corresponding years, months, or days.
}

