\name{compareIndivTrajectories}
\alias{compareIndivTrajectories}
\title{UNFINISHED FUNCTION!!}
\usage{
compareIndivTrajectories(dfr = NULL, indivConn = NULL, var = NULL,
  onlySuccessfulIndivs = TRUE, nurseryAlongshoreZones = 1:13,
  nurseryDepthZones = c("NurseryArea_000to050m", "NurseryArea_050to150m"),
  spawningAlongshoreZones = 1:12,
  spawningDepthZones = c("SpawningArea_300to600m"))
}
\description{

}

