\name{compareIndivTrajectories}
\alias{compareIndivTrajectories}
<<<<<<< HEAD
\title{UNFINISHED FUNCTION!!}
\usage{
compareIndivTrajectories(dfr = NULL, indivConn = NULL, var = NULL,
  onlySuccessfulIndivs = TRUE, nurseryAlongshoreZones = 1:13,
=======
\title{Compare individual trajectories}
\usage{
compareIndivTrajectories(dfr = NULL, indivConn = NULL, var = NULL,
  lhsTypeInfo = getLifeStageInfo.ATF(), onlySuccessfulIndivs = TRUE,
  nurseryAlongshoreZones = 1:13,
>>>>>>> 62a74cf16a4686e604456070f29fe0f6b451d39d
  nurseryDepthZones = c("NurseryArea_000to050m", "NurseryArea_050to150m"),
  spawningAlongshoreZones = 1:12,
  spawningDepthZones = c("SpawningArea_300to600m"))
}
\description{

}

