% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compareIndivTrajectories}
\alias{compareIndivTrajectories}
\title{Compare individual trajectories}
\usage{
compareIndivTrajectories(dfr = NULL, indivConn = NULL, var = NULL,
  lhsTypeInfo = getLifeStageInfo.ATF(), onlySuccessfulIndivs = TRUE,
  nurseryAlongshoreZones = 1:13,
  nurseryDepthZones = c("NurseryArea_000to050m", "NurseryArea_050to150m"),
  spawningAlongshoreZones = 1:12,
  spawningDepthZones = c("SpawningArea_300to600m"))
}
\description{

}

