\name{doAnalysis.FractionRecruitingBySource}
\alias{doAnalysis.FractionRecruitingBySource}
\title{Calculate time series of the total fraction recruiting by source.}
\usage{
doAnalysis.FractionRecruitingBySource(allRes)
}
\arguments{
  \item{allRes}{- results list object from
  calcConnectivityMatrices(...)}
}
\value{
matrix of fractions recruiting, with times as rows and
sources as columns
}
\description{
Function to calculate time series of the total fraction
recruiting by source.
}
\details{

}

