% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractIndivIDs.R
\name{extractIndivIDs}
\alias{extractIndivIDs}
\title{Function to extract unique IDs for individuals from a DisMELS model run.}
\usage{
extractIndivIDs(indivConn = NULL, lhsTypeInfo = NULL,
  onlySuccessful = FALSE, onlyUnsuccessful = FALSE,
  nurseryZones = c("NurseryArea_000to050m", "NurseryArea_050to150m"))
}
\arguments{
\item{indivConn}{- individual connectivity file (or dataframe)}

\item{lhsTypeInfo}{- life stage info list for IBM (req'd if onlySuccessful=TRUE or onlyUnsuccessful=TRUE)}

\item{onlySuccessful}{- flag (T/F) to pull out only successful individuals}

\item{onlyUnsuccessful}{- flag (T/F) to pull out only unsuccessful individuals}

\item{nurseryZones}{- vector of names of zones used as nursery areas in the IBM (req'd if onlySuccessful=TRUE)}
}
\value{
data frame with columns
ID, start_typeName,start_depthzone,start_alongshorezone,end_typeName,end_depthzone,end_alongshorezone
}
\description{

}
\details{
If onlySuccessful=TRUE, then it is assumed that 'successful' individuals are those which are in
one of the nursery areas in the final life stage.\cr
\cr
If onlyUnsuccessful=TRUE, then it is assumed that 'unsuccessful' individuals are those which
don't make it to the final life stage.
}

