% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractIndivs.MultipleRuns}
\alias{extractIndivs.MultipleRuns}
\title{Extracts and re-orders a series of DisMELS model results files by individual.}
\usage{
extractIndivs.MultipleRuns(YYYYs = NULL, MMs = NULL, DDs = NULL,
  lhsTypeInfo = NULL, lhsTypes = NULL, newResType = FALSE,
  onlySucccessful = TRUE, onlyUnsucccessful = FALSE,
  nurseryZones = c("NurseryArea_000to050m", "NurseryArea_050to150m"),
  inpDir.IndivConn = NULL, basename.IndivConn = "IndivConn",
  inpDir.Results = NULL, basename.Results = "Results", returnList = TRUE,
  writeOutput = TRUE, outDir = NULL, basename.Out = "SuccessfulIndivs")
}
\arguments{
\item{YYYYs}{- years to extract (can be NULL)}

\item{MMs}{- months to extract (can be NULL)}

\item{DDs}{- days to extract (can be NULL)}

\item{lhsTypeInfo}{- life stage info list for IBM}

\item{lhsTypes}{- lhs types to keep for output (NULL = all)}

\item{newResType}{- flag (T/F) indicating whether (T) or not (F) results are from a "new"-type IBM}

\item{nurseryZones}{- vector of nursery area names}

\item{inpDir.IndivConn}{- folder with input individual connectivity matrix files}

\item{basename.IndivConn}{- base name for individual connectivity matrix files}

\item{inpDir.Results}{- folder with input DisMELS model results files}

\item{writeOutput}{- flag (T/F) indicating whether output files should be written}

\item{outDir}{- output folder}

\item{basename.Out}{- base name for output files}

\item{onlySuccessful}{- flag (T/F) to extract only successful indivs (if TRUE)}

\item{onlyUnsuccessful}{- flag (T/F) to extract only unsuccessful indivs (if TRUE)}

\item{basename.Resuls}{- base name for DisMELS model results files}
}
\value{
no return value
}
\description{
Function to extract and re-order a series of DisMELS model results by individual.
}
\details{


This function assumes files are named something like 'basenameYYYYMMDD.csv'.
Note that at least one of YYYYs, MMs, and DDs must be a non-null vector.
}

