% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractIndivs}
\alias{extractIndivs}
\title{Function to extract tracks of individuals from a DisMELS model run.}
\usage{
extractIndivs(indivConn = NULL, results = NULL, lhsTypeInfo = NULL,
  lhsTypes = NULL, newResType = FALSE, onlySuccessful = FALSE,
  onlyUnsuccessful = FALSE, nurseryZones = c("NurseryArea_000to050m",
  "NurseryArea_050to150m"), returnList = TRUE, writeOutput = TRUE,
  outDir = "./", outBaseCSV = "allIndivs")
}
\arguments{
\item{indivConn}{- individual connectivity file (or dataframe)}

\item{results}{- DisMELS results file (or dataframe)}

\item{lhsTypeInfo}{- life stage info list for IBM}

\item{lhsTypes}{- vector of lhs types to process (NULL = all)}

\item{newResType}{- flag (T/F) indicating whether (TRUE) or not (FALSE) results are from a "new"-type IBM}

\item{onlySuccessful}{- flag (T/F) to extract only successful indivs (if TRUE)}

\item{onlyUnsuccessful}{- flag (T/F) to extract only unsuccessful indivs (if TRUE)}

\item{nurseryZones}{- vector of names of zones used as nursery areas in the IBM}

\item{returnList}{- flag (T/F) indicating whether a list should be returned (otherwise NULL--helps with memory)}

\item{writeOutput}{- flag (T/F) indicating whether output files should be written}

\item{outDir}{- folder for output}

\item{outBaseCSV}{- base name for output csv files}
}
\value{
list of dataframes with indiv results by lhs type name if returnList=TRUE, otherwise NULL.
}
\description{

}
\details{
If indivConn or results are NULL, the user will be prompted to select the
corresponding file(s) using a file chooser dialog.
}

