% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getLifeStageInfo.ATF.R
\name{getLifeStageInfo.ATF}
\alias{getLifeStageInfo.ATF}
\title{Get a list that defines life stage info for the DisMELS arrowtooth flounder IBM.
#}
\usage{
getLifeStageInfo.ATF()
}
\description{

}

