% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getStdVars}
\alias{getStdVars}
\title{Get standard variable names for DisMELS output.}
\usage{
getStdVars(newResType = TRUE)
}
\arguments{
\item{newResType}{- flag (T/F) indicating if results are based on the new or old DisMELS results format.}
}
\value{
character vector with names of standard variables
}
\description{
Function to get standard variable names for DisMELS output.
}
\details{

}

