% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStdVars.R
\name{getStdVars}
\alias{getStdVars}
\title{Get standard variable names for DisMELS output}
\usage{
getStdVars(newResType)
}
\arguments{
\item{newResType}{- flag ("NEW" or "OLD") indicating if results are based on the new or old DisMELS results format.}
}
\value{
data frame with columns for names of standard variables ('vars') and types ('types)
}
\description{
Function to get standard variable names for DisMELS output.
}
\details{
none
}

