% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConnectivityMatrixGG.R
\name{plotConnectivityMatrixGG}
\alias{plotConnectivityMatrixGG}
\title{Plot connectivity matrices using ggplot2.}
\usage{
plotConnectivityMatrixGG(dfr, x = "nursery area", y = "spawning area",
  z = "Pr(N|S)", xLab = x, yLab = y, zLab = z, xMin = NULL,
  xMax = NULL, yMin = NULL, yMax = NULL, zMin = NULL, zMax = NULL,
  facetWrap = NULL, facetGrid = NULL, ncol = 4, reverseX = FALSE,
  reverseY = FALSE, useGrad2 = FALSE)
}
\arguments{
\item{dfr}{- dataframe to plot}

\item{x}{- column name for x values}

\item{y}{- column name for y values}

\item{z}{- column name for z values}

\item{xLab}{- x axis label}

\item{yLab}{- y axis label}

\item{zLab}{- z label}

\item{xMin}{- x axis minimum}

\item{xMax}{- x axis maximum}

\item{yMin}{- y axis minimum}

\item{yMax}{- y axis maximum}

\item{zMin}{- z axis minimum}

\item{zMax}{- z axis maximum}

\item{facetWrap}{- formula for ggplot2::facet_wrap}

\item{facetGrid}{- formula for ggplot2::facet_grid}

\item{ncol}{- number of columns for ggplot2::facet_wrap}

\item{reverseX}{- flag to reverse x axis}

\item{reverseY}{- flag to reverse y axis}

\item{useGrad2}{- flag to use reshape2::scale_fill_gradient2 for fill axis}
}
\description{
Function to plot connectivity matrices using ggplot2.
}

